backend/requirements.txt
fastapi
uvicorn
SQLModel
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
